add: criado os get's e set's para os outros atributos e corrigidos os erros do codigo

+++ src/main/java/aula/ContaCorrente.java
@@ -9,23 +9,25 @@ package aula;
  *
  * @author samara
  */
-public class ContaCorrente extends Conta{
+public class ContaCorrente extends Conta {
+
     private double salario;
-    public ContaCorrente(int numero, double salario, Agencia agencia){
+
+    public ContaCorrente(int numero, double salario, Agencia agencia) {
         super(numero, agencia);
         this.salario = salario;
     }
 
     @Override
     public void imprimir() {
-        System.out.println("Seu salário é : "+this.salario);
+        System.out.println("Seu salário é : " + this.salario);
     }
-    
-    public double getSalario(){
+
+    public double getSalario() {
         return this.salario;
     }
-    
-    public void setSalario(int salario){
+
+    public void setSalario(double salario) {
         this.salario = salario;
-    }       
+    }
 }
